+++ CHANGELOG.md
@@ -1,3 +1,1 @@
-* Added option to switch between shadow and outline at experience level
-* The list of available versions has been expanded
-* Remove mandatory Cloth Config dependency again
+* Added compatibility with Appleskin